apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
sqlalchemy==2.0.36
psycopg==3.2.3
alembic==1.14.0
